Move SensorListFragment to sensors package
