+++ app/src/main/java/com/aconno/acnsensa/dagger/mainactivity/MainActivityComponent.kt
@@ -2,7 +2,7 @@ package com.aconno.acnsensa.dagger.mainactivity
 
 import com.aconno.acnsensa.dagger.application.AppComponent
 import com.aconno.acnsensa.ui.MainActivity
-import com.aconno.acnsensa.ui.SensorListFragment
+import com.aconno.acnsensa.ui.sensors.SensorListFragment
 import com.aconno.acnsensa.ui.beacons.BeaconListFragment
 import dagger.Component
